Reshape slot public API: type-only export + $getDOMSlot / $getDOMTextNode helpers
