+++ packages/lexical/src/nodes/__tests__/unit/LexicalElementNode.test.tsx
@@ -14,7 +14,6 @@ import {
   $getSelection,
   $isRangeSelection,
   createEditor,
-  ElementDOMSlot,
   ElementNode,
   LexicalEditor,
   LexicalNode,
@@ -30,7 +29,8 @@ import {
   $createTestElementNode,
   createTestEditor,
 } from '../../../__tests__/utils';
-import {indexPath, SerializedElementNode} from '../../LexicalElementNode';
+import {ElementDOMSlot, indexPath} from '../../../LexicalDOMSlot';
+import {SerializedElementNode} from '../../LexicalElementNode';
 
 describe('LexicalElementNode tests', () => {
   let container: HTMLElement;
